Update StraightTrackAlignmentDriver.java
Increase minimum number of hits in fit to 8
Recall that we have a maximum of 14 hits, but layers 1 & 2 are not illuminated by the field-off tracks. The top is missing all of layer 6, the bottom is missing layer4 axial.
Add cuts on reconstructed (x,y) position at wire. Allow events that pass these cuts to be skimmed.